Update: move HealthcheckupManagement to NurseView
