+++ WPF.SchoolMedicalManagementSystem/NurseView/HealthCheckupManagement.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="WPF.SchoolMedicalManagementSystem.ManagerView.HealthCheckupManagement"
+﻿<Window x:Class="WPF.SchoolMedicalManagementSystem.NurseView.HealthCheckupManagement"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:WPF.SchoolMedicalManagementSystem.ManagerView"
+        xmlns:local="clr-namespace:WPF.SchoolMedicalManagementSystem.NurseView"
         mc:Ignorable="d"
         Title="Health Checkups Management - School Medical System" Height="700" Width="1200"
         WindowStartupLocation="CenterScreen"
