+++ WPF.SchoolMedicalManagementSystem/ParentView/HealthCheckupResultManagement.xaml
@@ -9,6 +9,7 @@
         WindowStartupLocation="CenterScreen"
         Background="White">
     <Window.Resources>
+        <local:GenderToBackgroundConverter x:Key="GenderToBackgroundConverter"/>
         <!-- Modern Card Style -->
         <Style x:Key="ModernCardStyle" TargetType="Border">
             <Setter Property="Background" Value="White" />
